Split ASP.NET Core binder tests by type
